completed task

+++ main/routes.py
@@ -75,7 +75,7 @@ def search():
 @login_required
 def products():
     if session['role'] == 'customer':
-        products = Product.query.all()
+        products = Product.query.order_by(Product.price.asc()).all()
         return render_template('products.html', products = products)
     else:
         return redirect(url_for('seller_dashboard'))
@@ -90,13 +90,33 @@ def seller_products(id):
     else:
         return redirect(url_for('customer_home'))
 
-@app.route('/product/<int:id>')
+@app.route('/product/<int:id>', methods=['GET', 'POST'])
 @login_required
 def show_product(id):
     if session['role'] == 'customer':
         product = Product.query.filter_by(id=id).first()
         seller = User.query.filter_by(id = product.seller_id).first()
         form = AddToCart()
+        if form.validate_on_submit:
+            if form.quantity.data:
+                if form.quantity.data < product.quantity:
+                    cart = Cart.query.filter_by(user_id=current_user.id).all()
+                    alreadyPresent=False
+                    for c in cart:
+                        if product.id == c.product.id:
+                            c.quantity += form.quantity.data
+                            db.session.commit()
+                            alreadyPresent=True
+                            break
+                    if not alreadyPresent:
+                        cart1 = Cart(product = product, user_id=current_user.id, quantity=form.quantity.data)
+                        db.session.add(cart1)
+                        db.session.commit()
+                    flash('Product successfully added to Cart', 'check')
+                    return redirect(url_for('cart'))
+                else:
+                    flash('Currently {} pieces of this item are available'.format(product.quantity), 'error_outline')
+                    return redirect(url_for('customer_home'))
         return render_template('show_product.html', product = product, form = form, seller = seller)
     else:
         return redirect(url_for('seller_dashboard'))
@@ -107,16 +127,19 @@ def customer_home():
     if session['role'] == 'customer':
         mobiles = Product.query.filter_by(category='mobile').all()
         laptops = Product.query.filter_by(category='laptop').all()
-        form = AddToCart()
-        return render_template('customer_home.html', mobiles = mobiles[:4], laptops=laptops[:4], form = form)
+        return render_template('customer_home.html', mobiles = mobiles[:4], laptops=laptops[:4])
     else:
         return redirect(url_for('seller_dashboard'))
 
 @app.route('/customer/orders')
 @login_required
 def orders():
-    orders = Order.query.filter_by(user_id=current_user.id).all()
-    return render_template('orders.html', orders=orders)
+    if session['role'] == 'customer':
+        orders = Order.query.filter_by(user_id=current_user.id).order_by(Order.time.desc()).all()
+        return render_template('orders.html', orders=orders)
+    else:
+        flash('You cannot access this page', 'warning')
+        return redirect(url_for('seller_dashboard'))
 
 @app.route('/customer/cart')
 @login_required
@@ -131,100 +154,83 @@ def cart():
         flash('Only customer can access this page', 'warning')
         return redirect(url_for('seller_dashboard'))
 
-@app.route('/customer/cart/add/<int:id>', methods = ['GET', 'POST'])
-@login_required
-def add_to_cart(id):
-    product = Product.query.filter_by(id=id).first()
-    form = AddToCart()
-    if form.validate_on_submit:
-        if form.quantity.data < product.quantity:
-            cart = Cart.query.filter_by(user_id=current_user.id).all()
-            alreadyPresent=False
-            for c in cart:
-                if product.id == c.product.id:
-                    c.quantity += form.quantity.data
-                    db.session.commit()
-                    alreadyPresent=True
-                    break
-            if not alreadyPresent:
-                cart1 = Cart(product = product, user_id=current_user.id, quantity=form.quantity.data)
-                db.session.add(cart1)
-                db.session.commit()
-            flash('Product successfully added to Cart', 'check')
-            return redirect(url_for('cart'))
-        else:
-            flash('Currently {} pieces of this item are available'.format(product.quantity), 'error_outline')
-            return redirect(url_for('customer_home'))
-
 @app.route('/customer/cart/<string:operation>/<int:id>')
 @login_required
 def edit_cart(id, operation):
-    cart = Cart.query.filter_by(id=id).first()
-    if operation == 'increase':
-        if cart.quantity < cart.product.quantity:
-            cart.quantity += 1
-            db.session.commit()
-            flash('{} quantity updated'.format(cart.product.name), 'check')
-            return redirect(url_for('cart'))
+    if session['role'] == 'customer':
+        cart = Cart.query.filter_by(id=id).first()
+        if operation == 'increase':
+            if cart.quantity < cart.product.quantity:
+                cart.quantity += 1
+                db.session.commit()
+                flash('{} quantity updated'.format(cart.product.name), 'check')
+                return redirect(url_for('cart'))
+            else:
+                flash('Currently {} pieces of this item are available'.format(cart.product.quantity), 'warning')
+                return redirect(url_for('cart'))
+        elif operation == 'decrease':
+            if cart.quantity > 1:
+                cart.quantity -= 1
+                db.session.commit()
+                flash('{} quantity updated'.format(cart.product.name), 'check')
+                return redirect(url_for('cart'))
+            else:
+                flash('Quantity cannot be less than one.', 'warning')
+                return redirect(url_for('cart'))
         else:
-            flash('Currently {} pieces of this item are available'.format(cart.product.quantity), 'warning')
-            return redirect(url_for('cart'))
-    elif operation == 'decrease':
-        if cart.quantity > 1:
-            cart.quantity -= 1
+            db.session.delete(cart)
             db.session.commit()
-            flash('{} quantity updated'.format(cart.product.name), 'check')
-            return redirect(url_for('cart'))
-        else:
-            flash('Quantity cannot be less than one.', 'warning')
+            flash('Item removed successfully', 'check')
             return redirect(url_for('cart'))
     else:
-        db.session.delete(cart)
-        db.session.commit()
-        flash('Item removed successfully', 'check')
-        return redirect(url_for('cart'))
+        flash('Only customer can access this page', 'warning')
+        return redirect(url_for('seller_dashboard'))
 
 @app.route('/customer/checkout', methods=['GET', 'POST'])
 @login_required
 def checkout():
-    form = AddressForm()
-    cart = Cart.query.filter_by(user_id=current_user.id).all()
-    address = Address.query.filter_by(user_id=current_user.id).first()
-    total = 0
-    for item in cart:
-        total += item.product.price*item.quantity
-    if form.validate_on_submit():
-        if address is None:
-            address = Address(addressLine1 = form.addressLine1.data,
-                            addressLine2 = form.addressLine2.data,
-                            pincode = form.pincode.data,
-                            city = form.city.data,
-                            state = form.state.data,
-                            mobile = form.mobile.data,
-                            customer = current_user)
-            db.session.add(address)
-            db.session.commit()
+    if session['role'] == 'customer':
+        form = AddressForm()
+        cart = Cart.query.filter_by(user_id=current_user.id).all()
+        address = Address.query.filter_by(user_id=current_user.id).first()
+        total = 0
         for item in cart:
-            order = Order(product = item.product, user_id = current_user.id, seller_id=item.product.seller_id, quantity=item.quantity, total = total)
-            db.session.add(order)
-            db.session.commit()
-        for item in cart:
-            product = Product.query.filter_by(id=item.product.id).first()
-            product.quantity -= item.quantity
-            db.session.commit()
-        for item in cart:
-            Cart.query.filter_by(id=item.id).delete()
-            db.session.commit()
-        flash('Order Placed successfully', 'check')
-        return redirect(url_for('orders'))
-    if request.method == 'GET' and address is not None:
-        form.addressLine1.data = address.addressLine1
-        form.addressLine2.data = address.addressLine2
-        form.pincode.data = address.pincode
-        form.city.data = address.city
-        form.state.data = address.state
-        form.mobile.data = address.mobile
-    return render_template('checkout.html', form=form, total=total, cart_items=cart)
+            total += item.product.price*item.quantity
+        if form.validate_on_submit():
+            if address is None:
+                address = Address(addressLine1 = form.addressLine1.data,
+                                addressLine2 = form.addressLine2.data,
+                                pincode = form.pincode.data,
+                                city = form.city.data,
+                                state = form.state.data,
+                                mobile = form.mobile.data,
+                                customer = current_user)
+                db.session.add(address)
+                db.session.commit()
+            for item in cart:
+                order = Order(product = item.product, user_id = current_user.id, seller_id=item.product.seller_id, quantity=item.quantity, total = total)
+                db.session.add(order)
+                db.session.commit()
+            for item in cart:
+                product = Product.query.filter_by(id=item.product.id).first()
+                product.quantity -= item.quantity
+                db.session.commit()
+            for item in cart:
+                Cart.query.filter_by(id=item.id).delete()
+                db.session.commit()
+            flash('Order Placed successfully', 'check')
+            return redirect(url_for('orders'))
+        if request.method == 'GET' and address is not None:
+            form.addressLine1.data = address.addressLine1
+            form.addressLine2.data = address.addressLine2
+            form.pincode.data = address.pincode
+            form.city.data = address.city
+            form.state.data = address.state
+            form.mobile.data = address.mobile
+        return render_template('checkout.html', form=form, total=total, cart_items=cart)
+    else:
+        flash('Only customer can access this page', 'warning')
+        return redirect(url_for('seller_dashboard'))
 
 @app.route('/dashboard/seller')
 @login_required
@@ -234,40 +240,55 @@ def seller_dashboard():
         no_of_products = Product.query.filter_by(seller=current_user).count()
         no_of_orders = Order.query.filter_by(seller_id = current_user.id).count()
         graph=[]
-        prev = Order.query.group_by(sa.func.strftime('%d')).count()
-        graph.append(prev)
+        j = Order.query.filter_by(seller_id=current_user.id).first().time
+        j = int(j.strftime('%d'))
+        for i in range(7):
+            prev = Order.query.filter(Order.time > datetime(2020, 9, j), Order.time < datetime(2020, 9, j + 1)).count()
+            if prev == 0 and j == int(datetime.today().strftime('%d')) + 1:
+                break
+            j += 1
+            graph.append(prev)
         return render_template('seller_dashboard.html', products=products, no_of_products = no_of_products, no_of_orders=no_of_orders, graph=graph)
     else:
+        flash('Only Seller can access this page', 'warning')
         return redirect(url_for('customer_home'))
 
 @app.route('/dashboard/seller/history')
 @login_required
 def seller_history():
-    seller_orders = Order.query.filter_by(seller_id = current_user.id).all()
-    customers = []
-    for seller_order in seller_orders:
-        customer_id = seller_order.user_id
-        customer = User.query.filter_by(id=customer_id).first()
-        customers.append(customer)
-    return render_template('seller_history.html',context = zip(seller_orders, customers))
+    if session['role'] == 'seller':
+        seller_orders = Order.query.filter_by(seller_id = current_user.id).order_by(Order.time.desc()).all()
+        customers = []
+        for seller_order in seller_orders:
+            customer_id = seller_order.user_id
+            customer = User.query.filter_by(id=customer_id).first()
+            customers.append(customer)
+        return render_template('seller_history.html',context = zip(seller_orders, customers))
+    else:
+        flash('Only Seller can access this page', 'warning')
+        return redirect(url_for('customer_home'))
 
 @app.route('/dashboard/seller/new_product', methods=['POST', 'GET'])
 @login_required
 def new_product():
-    form = ProductForm()
-    if form.validate_on_submit():
-        product = Product(name = form.name.data,
-                          description = form.description.data,
-                          category = form.category.data,
-                          price = form.price.data, 
-                          quantity = form.quantity.data,
-                          photo_name = save_picture(form.photo.data),
-                          seller = current_user)
-        db.session.add(product)
-        db.session.commit()
-        flash('Product has been added successfully!', 'check')
-        return redirect(url_for('seller_dashboard'))
-    return render_template('new_product.html', legend = "Add Product", form=form)
+    if session['role'] == 'seller':
+        form = ProductForm()
+        if form.validate_on_submit():
+            product = Product(name = form.name.data,
+                            description = form.description.data,
+                            category = form.category.data,
+                            price = form.price.data, 
+                            quantity = form.quantity.data,
+                            photo_name = save_picture(form.photo.data),
+                            seller = current_user)
+            db.session.add(product)
+            db.session.commit()
+            flash('Product has been added successfully!', 'check')
+            return redirect(url_for('seller_dashboard'))
+        return render_template('new_product.html', legend = "Add Product", form=form)
+    else:
+        flash('Only Seller can access this page', 'warning')
+        return redirect(url_for('customer_home'))
 
 def save_picture(form_picture):
     if form_picture:
@@ -286,29 +307,33 @@ def save_picture(form_picture):
 @app.route('/seller/product/<int:id>/update', methods=['GET', 'POST'])
 @login_required
 def update_product(id):
-    product = Product.query.get_or_404(id)
-    if product.seller != current_user:
-        abort(403)
-    form = ProductForm()
-    if form.validate_on_submit():
-        product.name = form.name.data
-        product.description = form.description.data
-        product.category = form.category.data
-        product.price = form.price.data
-        product.quantity = form.quantity.data
-        if product.photo_name is None:
-            product.photo_name = save_picture(form.photo.data)
-        elif form.photo.data:
-            pathp = app.root_path + '\static\product_pics\{}'.format(product.photo_name)
-            os.remove(pathp)
-            product.photo_name = save_picture(form.photo.data)
-        db.session.commit()
-        flash('Product has been updated!', 'check')
-        return redirect(url_for('seller_dashboard'))
-    elif request.method == 'GET':
-        form.name.data = product.name
-        form.category.data = product.category
-        form.description.data = product.description
-        form.price.data = product.price
-        form.quantity.data = product.quantity
-    return render_template('new_product.html', legend='Update Product', form=form)
+    if session['role'] == 'seller':
+        product = Product.query.get_or_404(id)
+        if product.seller != current_user:
+            abort(403)
+        form = ProductForm()
+        if form.validate_on_submit():
+            product.name = form.name.data
+            product.description = form.description.data
+            product.category = form.category.data
+            product.price = form.price.data
+            product.quantity = form.quantity.data
+            if product.photo_name is None:
+                product.photo_name = save_picture(form.photo.data)
+            elif form.photo.data:
+                pathp = app.root_path + '\static\product_pics\{}'.format(product.photo_name)
+                os.remove(pathp)
+                product.photo_name = save_picture(form.photo.data)
+            db.session.commit()
+            flash('Product has been updated!', 'check')
+            return redirect(url_for('seller_dashboard'))
+        elif request.method == 'GET':
+            form.name.data = product.name
+            form.category.data = product.category
+            form.description.data = product.description
+            form.price.data = product.price
+            form.quantity.data = product.quantity
+        return render_template('new_product.html', legend='Update Product', form=form)
+    else:
+        flash('Only Seller can access this page', 'warning')
+        return redirect(url_for('customer_home'))
